@@ -1,13 +1,13 @@
 package fr.free.edgar35740.camerash
 
-import android.support.v4.content.ContextCompat
+import androidx.core.content.ContextCompat
 
 import com.jjoe64.graphview.GraphView
 import com.jjoe64.graphview.series.LineGraphSeries
 
 class ChartController(private val view: GraphView, range: Int) {
 
-    private val data = Array(range, { i -> DataPoint(i, 0) })
+    private val data = Array(range) { i -> DataPoint(i, 0) }
     private val chartData = LineGraphSeries(data)
     private var max = 0f
 
@@ -18,7 +18,7 @@ class ChartController(private val view: GraphView, range: Int) {
     fun pushNumber(number: Int) {
         if (number !in 0..data.size) return
         data[number]++
-        max = Math.max(max, data[number].y.toFloat())
+        max = max.coerceAtLeast(data[number].y.toFloat())
         chartData.isDrawDataPoints = true
         chartData.setAnimated(true)
         chartData.color = ContextCompat.getColor(view.context, R.color.colorAccent)
